Questions
I created a a post and get request for the questions table. Going to start with the update request now. In the post request, the user can post a new question in the table into a specific column of their choosing. The get request allows the user to search for all the question, questions with its business type(new or existing), by it's heading (business_description, etc.), by it's id or by a combination of them all.